requirements.txt
# ==================================================
# ✅ Core Dependencies (الاعتمادات الأساسية)
# ==================================================
streamlit==1.46.0              # واجهة المستخدم
pandas==2.3.1                  # معالجة البيانات
pytz==2024.1                   # المناطق الزمنية
pydantic==2.8.2                # التحقق من المدخلات والنماذج
python-dotenv==1.0.1           # تحميل متغيرات البيئة

# ==================================================
# ✅ Web Scraping
# ==================================================
beautifulsoup4==4.12.3         # تحليل HTML
lxml==5.2.2                    # محرك HTML/ XML سريع (متوافق مع Python 3.11)
greenlet==3.0.3                    
playwright==1.45.0             # أداة التصفح الآلي
redis==5.0.7                   # التخزين المؤقت

# ==================================================
# ✅ Plotting & Visualization
# ==================================================
plotly==6.2.0                  # الرسومات التفاعلية

# ==================================================
# ✅ Error Reporting
# ==================================================
sentry-sdk==2.8.0              # إرسال الأخطاء إلى Sentry

# ==================================================
# ✅ Database
# ==================================================
psycopg2-binary==2.9.9         # PostgreSQL driver
SQLAlchemy==2.0.31             # ORM وإدارة قواعد البيانات

# ==================================================
# ✅ Testing
# ==================================================
pytest==8.3.2
pytest-mock==3.14.0
